requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
Werkzeug<3.0.0
python-dotenv==1.1.1